requirements.txt
django
psycopg2-binary
djongo
django-bootstrap-v5
pymongo
sshtunnel